sources/requirements.txt
requests-html
cfscrape
mechanize
ruamel.yaml
sh
image_slicer